requirements.txt
opencv-python
numpy
mss
easyocr
torch
neat-python
pyyaml
